Use better logical child tracking
For now there are internal methods we can call for adding/removing logical children without having to maintain a list ourselves.  This moves the logic out into extension methods which currently uses reflection to get either the internal method, or if that's not found, on newer .NET 8 previews/rc's the method is public so we can safeguard to get that via reflection until a newer sdk is released where we can compile against the upcoming change.